Add managment command to re-localize timezones
In my old deployment guide[^0] I propsed a setup without Timezone
support. But when importing the data in a itmezone-aware instalation,
all of the dates are bad. Therefore this command can convert naive
timestaps to aware timestamps.

Usage:

```
time ./manage.py fixdatetime Europe/Berlin -v2
```

In this case all demo dates in the database would be treated as
'Europe/Berlin' originially and then saved back to the database.

[^0]: https://docs.demostat.de/de/master/deploy.html